add: CardHover component for render information
